+++ CHANGELOG.md
@@ -12,6 +12,7 @@ adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 - Added a `sidebar` slot to `<manifold-marketplace>` (#612)
 - Exposed `ensureAuthToken` function for making authenticated GraphQL calls directly from platforms
   (#576)
+- Added `ownerId` prop to `<manifold-data-provision-button>`
 
 ### Removed